+++ stable_ssl/__init__.py
@@ -17,7 +17,12 @@ from .base import BaseTrainer
 from .config import instanciate_config
 from .losses import BarlowTwinsLoss, NegativeCosineSimilarity, NTXEntLoss, VICRegLoss
 from .modules import load_backbone
-from .trainers import JointEmbeddingTrainer, SelfDistillationTrainer, SupervisedTrainer
+from .trainers import (
+    JointEmbeddingPredictiveTrainer,
+    JointEmbeddingTrainer,
+    SelfDistillationTrainer,
+    SupervisedTrainer,
+)
 
 __all__ = [
     "__title__",
@@ -30,10 +35,12 @@ __all__ = [
     "BaseTrainer",
     "SupervisedTrainer",
     "JointEmbeddingTrainer",
+    "JointEmbeddingPredictiveTrainer",
     "SelfDistillationTrainer",
     "NTXEntLoss",
     "VICRegLoss",
     "BarlowTwinsLoss",
     "NegativeCosineSimilarity",
     "instanciate_config",
+    "random_mask",
 ]